feat: implement favorite screen, and add postpone, etc for the ui

+++ app/src/main/java/com/raihan/githubapp/ui/fragment/favorite_page/FavoriteViewModel.kt
@@ -1,6 +1,8 @@
 package com.raihan.githubapp.ui.fragment.favorite_page
 
 import android.app.Application
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.raihan.githubapp.data.local.UserFavoriteRepository
@@ -10,16 +12,25 @@ import kotlinx.coroutines.launch
 class FavoriteViewModel(application: Application) : ViewModel() {
 	private val mFavoriteRepository: UserFavoriteRepository =
 		UserFavoriteRepository(application)
+	private var _isLoading = MutableLiveData(true)
+	val isLoading: LiveData<Boolean> = _isLoading
 
-	fun insert(user: UserFavoriteEntity) {
-		viewModelScope.launch {
-			mFavoriteRepository.insert(user)
-		}
+	fun getAllFavorites(): LiveData<List<UserFavoriteEntity>> {
+		_isLoading.value = true
+		val user = mFavoriteRepository.getAllFavUser()
+		_isLoading.value = false
+		return user
 	}
 
-	fun delete(user: UserFavoriteEntity) {
+	fun toggleFavorite(user: UserFavoriteEntity, isFavorite: Boolean) {
+		_isLoading.value = true
 		viewModelScope.launch {
-			mFavoriteRepository.delete(user)
+			if (!isFavorite) {
+				mFavoriteRepository.insert(user)
+			} else {
+				mFavoriteRepository.delete(user)
+			}
+			_isLoading.value = false
 		}
 	}
